refactor: move getGoalDetails
